pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "distributed-hyperparam-search"
version = "0.1.0"
description = "Distributed hyperparameter search for machine learning pipelines"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["machine-learning", "hyperparameter-search", "distributed"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "redis>=4.0.0",
    "pyyaml>=6.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
ml = [
    "xgboost>=1.5.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
dhs-manager = "distributed_hyperparam_search.main:main"
dhs-worker = "distributed_hyperparam_search.worker:main"

[project.urls]
Homepage = "https://github.com/yourusername/distributed-hyperparam-search"
Documentation = "https://github.com/yourusername/distributed-hyperparam-search#readme"
Repository = "https://github.com/yourusername/distributed-hyperparam-search.git"
Issues = "https://github.com/yourusername/distributed-hyperparam-search/issues"

# Hatchling will auto-discover the package

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=distributed_hyperparam_search"

# UV specific settings
[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]